Fix glob.match built-in calls to provide delimiter
The glob.match built-in function was originally defined to default the
delimiter to "." when the parameter was an empty array, however,
someone recently discovered that this was not happening. Unfortunately
the fix is not backwards compatible and we felt it was better to
resolve the issue now rather than make the change in a backwards
compatible way (e.g., by deprecating glob.match, introducing a new
function, etc.) The fix is contained in OPA v0.17 and newer.